Move test framework into sharable internal package
